Task 2 and Task 3
I added a listener handler for the convert method. I updated the conditions in the changeLabels() method and the calcMileage method. I make sure to get numeric values only by adding a condition statement in the convert method. I made another if / else statement to make sure the values are converted when the cbo value is changed and it also updates in the calcMileage method. Ensured that the code works as described before submitting it to Gradescope.